pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "job-application-automation"
version = "1.0.0"
description = "AI-powered job application email automation with SMTP delivery and CSV tracking"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "google-generativeai>=0.7.0",
    "openai>=1.30.0",
    "rich>=13.0.0",
    "cyclonedx-bom>=4.0.0",
]

[project.scripts]
job-apply = "job_applier:main"

[tool.setuptools.packages.find]
where = ["."]

[project.urls]
Repository = "https://github.com/youruser/job-application-automation"
